requirements.txt
mysql-connector-python
SpeechRecognition
pygame
requests
cohere
pyaudio